requirements.txt
django
requests
django-ckeditor-5
django-hitcount
django-import-export